Extract WallPositionsGenerator to own header
