+++ src/game/SimpleMap.hpp
@@ -2,23 +2,10 @@
 #include <vector>
 
 #include "ExplosionRange.hpp"
+#include "WallPositionsGenerator.hpp"
 
 #include "glm/glm.hpp"
 #include "physics/PhysicsEngine.hpp"
-
-class WallPositionsGenerator
-{
-public:
-    using BaseType = int;
-    using PositionInSpace = std::pair<BaseType, BaseType>;
-    using WallSize = std::pair<BaseType, BaseType>;
-    using BoundarySize = int;
-    using Wall = std::pair<PositionInSpace, WallSize>;
-    using Walls = std::vector<Wall>;
-
-    virtual ~WallPositionsGenerator() = default;
-    virtual Walls generateBoundaryWalls(BoundarySize) const = 0;
-};
 
 class SimpleMap
 {
