+++ src/game/WallPositionsGenerator.hpp
@@ -1,0 +1,15 @@
+#pragma once
+
+class WallPositionsGenerator
+{
+public:
+    using BaseType = int;
+    using PositionInSpace = std::pair<BaseType, BaseType>;
+    using WallSize = std::pair<BaseType, BaseType>;
+    using BoundarySize = int;
+    using Wall = std::pair<PositionInSpace, WallSize>;
+    using Walls = std::vector<Wall>;
+
+    virtual ~WallPositionsGenerator() = default;
+    virtual Walls generateBoundaryWalls(BoundarySize) const = 0;
+};
